# Input Rewrite