--- a/source/backend/ClientPrefs.hx
+++ b/source/backend/ClientPrefs.hx
@@ -64,6 +64,7 @@ class SaveVariables {
 		'instakill' => false,
 		'practice' => false,
 		'botplay' => false,
+		'guitarherosustains' => false,
 		'opponentplay' => false
 	];