# Input V3 or some shit lmfao

---

Did a small rewrite on the Input System, it behaves pretty much the same, however you now have an Optional Gameplay Changer that enables Guitar Hero-like sustains (shoutouts to Kade Engine users!!!)

---

## Changes:

---

- Input Key Press function is much faster now, with better hit detection along with stacked notes (jacks) detection
- Rewrote how Hold Notes behave as an optional gameplay changer, whenever you miss a sustain with said changer enabled, the game shouldn't let you hold the sustain again, nor should it let you miss
- "Misses" label on Score Bar gets hidden if `Instakill on Miss` is enabled
- Misc Changes to make the code slightly cleaner