@crowplexus 2023-10-14
changed jack detection a lil bit, can any of yall test it?

specifically with songs that have jumps/doubles or triples, and/or have jacks