@TheRealJake12 2023-10-14
> good job bro. I'd actually like this change a lot.

I really dislike normal psych input. I hope this makes it feel better (I haven't tried it yet)