@crowplexus 2023-10-14
anyone willing to contribute to this, I need help with sustain inputs, basically when you release a sustain after successfully grabbing it, you miss like if it was still multiple notes

I only need a way of fixing that behavior