@TheRealJake12 2023-10-14
> anyone willing to contribute to this, I need help with sustain inputs, basically on guitar hero sustains, when you release a sustain after successfully grabbing it, you miss like if it was still multiple notes
> 
> I only need a way of fixing that behavior

steal kade engine's code. sustainActive to be specific in both PlayState.hx and Note.hx