@crowplexus 2023-10-16
> 
> Guess this is my cue.
> 
> I'd say it's much more consistent than before as it was much easier to SFC this, tweaked a chart I did to test jacks and chords. May be placebo? Idk, but I managed to SFC 4 times in a row, although this chart is pretty easy. I could import some osu!mania 5.5 star maps and higher aswell to test out if you want
> jack.and.chord.test.200bpm.mp4

thats a very good sign actually, my goal with this was to make the input function faster in general while also adding the GH/Kade Engine sustains as an optional bonus so yeah very nice to know