@Rom4chek 2023-10-16
Hey i found bug when without ghost tapping when you just pressing buttons, game crashes

smth wrong in this line ig idk

![image](https://github.com/ShadowMario/FNF-PsychEngine/assets/97591537/307db17c-99de-4350-b999-d56dc0bc60d3)