@crowplexus 2023-10-24
update to this: as a complementary update, I thought of replacing the etterna timings with the base game timings, adding the "shit" rating as a customizable one in the options menu, and let users change timings based on the worst timing's max

so you wouldn't be limited to 45ms in sick, for example, you would instead be limited to 166.67ms, which is the default timing for shit in base game

just an idea, I'm likely not gonna do it in this PR specifically