@crowplexus 2023-10-24
> 
> Personally as a rhythm game nut I think that would be a bad idea. The timings are already on the extremely lenient side. 45ms is A LOT

the purpose of this change would be mainly making timings extremely customizable, the essential "revert" of timings would be there because currently in psych, it is nearly impossible to get a Boo/Shit, this is because the timing is 180.0ms, which safeFrames can BARELY reach

| Judgement | Current Timing | New Timing |
| --------- | ------ | ---------- |
| SICK | 45.0ms | 33.33ms |
| GOOD | 90.0ms | 91.67ms |
| BAD | 135.0ms | 133.33ms |
| SHIT | 180.0ms | 166.67ms |

by changing timings to this, not only Boos/Shits would be able to be hit in normal gameplay, but I'd also change how the options themselves work so you can actually go back to the etterna timings if you want, or have completely custom ones which also allow for decimals to be used

and also it kinda bothers me that Boos can't have custom timings for no reason lol.